Pull code from utils into tar
